move reattach config
