+++ client/pluginmanager/drivermanager/state/state.go
@@ -1,11 +1,11 @@
 package state
 
-import "github.com/hashicorp/nomad/plugins/shared"
+import pstructs "github.com/hashicorp/nomad/plugins/shared/structs"
 
 // PluginState is used to store the driver managers state across restarts of the
 // agent
 type PluginState struct {
 	// ReattachConfigs are the set of reattach configs for plugin's launched by
 	// the driver manager
-	ReattachConfigs map[string]*shared.ReattachConfig
+	ReattachConfigs map[string]*pstructs.ReattachConfig
 }
