+++ plugins/shared/structs/plugin_reattach_config.go
@@ -1,4 +1,4 @@
-package shared
+package structs
 
 import (
 	"fmt"
